[log] Move NewXXLogger to log/factory.go
- remove usage of log package in util package to avoid future cycle
import, and there should be no usage of log inside util function, err
should be used, and most time they should be too small to have error
that should be logged
- deprecate old log/entry.go, it was created when we were still trying
to make a tree structure using log entry (which actually acted like our
current logger)
