+++ runner/context.go
@@ -1,5 +1,6 @@
 package runner
 
+// FIXME: this is never used, and it may not be a good idea to use context, since there is context package
 // Context describes how the command should be run
 // i.e. use shell instead of using os/exec,
 // use self defined logic instead of lookup an executable
