+++ util/pkg.go
@@ -1,9 +1,2 @@
 // Package util contains helper for test
 package util
-
-import (
-	"github.com/dyweb/gommon/util/logutil"
-)
-
-// log entry for util package
-var log = logutil.Logger.RegisterPkg()
